+++ analysis_options.yaml
@@ -1,0 +1,4 @@
+# include: package:flutter_lints/flutter.yaml
+
+# Additional information about this file can be found at
+# https://dart.dev/guides/language/analysis-options
