--- android/gradle.properties
@@ -1,1 +1,0 @@
-org.gradle.jvmargs=-Xmx1536M
