--- android/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Apr 15 11:28:54 COT 2019
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.1.1-all.zip
