+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.rioapp.demo.imeiplugin">
+  package="com.rioapp.demo.imeiplugin.imei_plugin">
 
-    <uses-permission android:name="android.permission.READ_PHONE_STATE" />
+  <uses-permission android:name="android.permission.READ_PHONE_STATE" />
+  <uses-permission android:name="android.permission.READ_PRIVILEGED_PHONE_STATE" />
 
 </manifest>
