+++ example/.flutter-plugins-dependencies
@@ -1,0 +1,1 @@
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"imei_plugin","path":"/Users/pablo384/projects/orkapi/imei_plugin/","dependencies":[]}],"android":[{"name":"imei_plugin","path":"/Users/pablo384/projects/orkapi/imei_plugin/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"imei_plugin","dependencies":[]}],"date_created":"2022-03-11 12:10:28.108249","version":"2.8.0"}
