--- example/android/app/src/main/java/com/rioapp/demo/imeipluginexample/MainActivity.java
@@ -1,13 +1,0 @@
-package com.rioapp.demo.imeipluginexample;
-
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
-
-public class MainActivity extends FlutterActivity {
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
-}
