+++ example/android/app/src/main/kotlin/com/rioapp/demo/imeiplugin/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.rioapp.demo.imeiplugin.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}
