+++ pubspec.yaml
@@ -1,7 +1,6 @@
 name: imei_plugin
 description: get real IMEI for android devices with validate permission on ejecution time and get unique id on ios An alphanumeric string that uniquely identifies a device to the app’s vendor.
-version: 2.0.0
-author: Kevin Caicedo <ing.sys.kevincaicedo@gmail.com>
+version: 2.1.0
 homepage: https://github.com/kevincaicedo/imei_plugin
 
 environment:
@@ -21,7 +20,7 @@ flutter:
   # be modified. They are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
   plugin:
-    androidPackage: com.rioapp.demo.imeiplugin
+    androidPackage: com.rioapp.demo.imeiplugin.imei_plugin
     pluginClass: ImeiPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
